LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-svg-script-expected.txt

CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"

PASS Assign String to SVGScriptElement.innerHTML.
PASS Assign TrustedHTML to SVGScriptElement.innerHTML.
PASS Assign TrustedHTML to SVGScriptElement.innerHTML and execute it.
PASS Modify SVGScriptElement via DOM manipulation.
